Separate files for several view model contexts
