+++ catowseriOS/CottonViewModels/WebViewModel/WebViewContext.swift
@@ -17,19 +17,20 @@ import CoreBrowser
  https://swiftrocks.com/whats-any-understanding-type-erasure-in-swift
  */
 
-/// Web view context should carry some data or dependencies which can't be stored as a state and always are present.
-/// protocol with async functions which do not belong to specific actor.
+/// Web view context should carry some data or dependencies which can't be stored as a state
+/// and always are present. Protocol with async functions which do not belong to specific actor.
 ///
 /// Can be sendable because doesn't store anything mutable.
 public protocol WebViewContext: Sendable {
-    /// Plugins are optional because there is possibility that js files are not present or plugins delegates are not set
+    /// Plugins are optional because there is possibility that js files are not present
+    /// or plugins delegates are not set
     var  pluginsSource: any JSPluginsSource { get }
     /// Hides app specific implementation for host check
     func nativeApp(for host: CottonBase.Host) -> String?
     /// Hides app specific feature for JS value
     func isJavaScriptEnabled() async -> Bool
     /// Hides app specific feature value for DNS over HTTPs
-    func isDohEnabled() async -> Bool
+    var isDohEnabled: Bool { get async }
     /// Hides app specific feature value for Native app redirects
     func allowNativeAppRedirects() async -> Bool
     /// Wrapper for feature value from specific app
